+++ client/src/app/modules/dashboard/modules/module-instance/components/filter-dialog/filter-dialog.component.ts
@@ -5,6 +5,7 @@ import {State, Parser, safeEval} from '@jaspero/form-builder';
 import {FilterMethod} from '../../../../../../shared/enums/filter-method.enum';
 import {FilterModule, FilterModuleDefinition} from '../../../../../../shared/interfaces/filter-module.interface';
 import {WhereFilter} from '../../../../../../shared/interfaces/where-filter.interface';
+import {InstanceOverviewContextService} from '../../services/instance-overview-context.service';
 
 @Component({
   selector: 'jms-filter-dialog',
@@ -15,7 +16,10 @@ import {WhereFilter} from '../../../../../../shared/interfaces/where-filter.inte
 export class FilterDialogComponent {
   constructor(
     @Inject(MAT_DIALOG_DATA)
-    public data: FilterModule,
+    public data: {
+      module: FilterModule,
+      ioc: InstanceOverviewContextService
+    },
     private dialogRef: MatDialogRef<FilterDialogComponent>
   ) {}
 
@@ -32,20 +36,40 @@ export class FilterDialogComponent {
     for (const key in data) {
       if (data.hasOwnProperty(key)) {
 
-        const definition = ((this.data.definitions || {})[key] || {}) as FilterModuleDefinition;
+        const definition = ((this.data.module.definitions || {})[key] || {}) as FilterModuleDefinition;
+        const value = data[key];
 
-        toSend.push({
-          key,
-          value: data[key],
-          operator: definition.filterMethod || FilterMethod.Equal,
-          ...definition.filterLabel && {label: definition.filterLabel},
-          ...this.data.persist && {persist: this.data.persist}
-        });
+        let displayValue: any = value;
+
+        if (
+          value !== undefined &&
+          value !== null &&
+          !Number.isNaN(value) &&
+          value !== ''
+        ) {
+          if (definition.filterValuePipe) {
+            displayValue = this.data.ioc.columnPipe.transform(
+              displayValue,
+              definition.filterValuePipe,
+              definition.filterValuePipeArguments,
+              {[definition.filterKey || key]: displayValue}
+            )
+          }
+
+          toSend.push({
+            value,
+            displayValue,
+            key: definition.filterKey || key,
+            operator: definition.filterMethod || FilterMethod.Equal,
+            ...definition.filterLabel && {label: definition.filterLabel},
+            ...this.data.module.persist && {persist: this.data.module.persist}
+          });
+        }
       }
     }
 
-    if (this.data.formatOnSubmit) {
-      const formatOnSubmit = safeEval(this.data.formatOnSubmit);
+    if (this.data.module.formatOnSubmit) {
+      const formatOnSubmit = safeEval(this.data.module.formatOnSubmit);
 
       if (formatOnSubmit) {
         toSend = formatOnSubmit(toSend);